requirements.txt
flask==3.0.3
gunicorn==22.0.0
replit-object-storage>=1.0.0
